Connect slots to configLoaded.
All the necessary hardware calls are done in the thread.  This avoids
freezing the GUI while the configuration is being loaded.

Starting the polling timer is also connected to configLoaded.

The widgetmappers are also now connected to configLoaded.  This fixes a
bug where the coal to mapper.toFirst() would be executed on an empty
model.  The results was that the user had to select a row in the table
so that the mappers actually work.