Relicense to MIT
I sold this project to my current employer and I am not able to maintain
it anymore.  I just wish it is useful to somebody.  I am still using the
code here without the Qt frontent and with the backend ported to
asyncio.  I am still happy with the API and I have not found any major
bug.  The MIT is a lot easier to handle than the GPL so that I now
relicense this project to The MIT License (MIT).